+++ app/Http/Controllers/HomeController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\Notes;
-use App\Models\Tag;
 use App\Models\Setting;
 use App\Models\User;
 use DB;
@@ -11,7 +10,6 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Services\SettingService;
-use Tiptap;
 
 class HomeController extends Controller
 {
